Adjust test helpers to new Go version. Remove tests and helpers to not supported versions.
