+++ tests/e2e/helper.go
@@ -1,6 +1,7 @@
 package e2e
 
 import (
+	"context"
 	crypto_rand "crypto/rand"
 	"encoding/binary"
 	"errors"
@@ -18,6 +19,18 @@ import (
 	pubnub "github.com/pubnub/go/v8"
 	"github.com/stretchr/testify/assert"
 )
+
+const (
+	ERR_CONTEXT_CANCELLED = "context canceled"
+	ERR_CONTEXT_DEADLINE  = "context deadline exceeded"
+)
+
+func contextWithTimeout(parent context.Context,
+	timeout time.Duration) (context.Context, context.CancelFunc) {
+	return context.WithTimeout(parent, timeout)
+}
+
+var backgroundContext = context.Background()
 
 var enableDebuggingInTests = false
